proveedores_ms/requirements.txt
django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==2.1.0
gunicorn==20.1.0
djongo==1.3.6
django-heroku==0.3.1